render less stuff
Previously, we took the naive approach: render the entire memory map.
This ended up being about 200k widgets for the starting memory map/fold
situation (3 widgets per byte*65536 bytes), and brick just wasn't
prepared for that.

Now we explicitly track what's on-screen, and only render that,
resulting in something in the 100s of widgets instead (<=6
widgets/line), which works much more smoothly.

As a side effect, we are now doing a fancy thing that we couldn't
reasonably have done with the old design: indicating if part of a fold
summary is off-screen.